+++ lib/seq_ext.mli
@@ -1,5 +1,0 @@
-val first : 'a Seq.t -> 'a
-
-val rest : 'a Seq.t -> 'a Seq.t
-
-val take_list : int -> 'a Seq.t -> 'a list