+++ test/test_basics.ml
@@ -1,36 +1,26 @@
-let take_list = Zar.take_list
-let first = Zar.first
-let rest = Zar.rest
-
-let rec string_of_list (sep : string) (to_string : 'a -> string) = function
-  | [] -> ""
-  | x :: xs -> to_string x ^ sep ^ string_of_list sep to_string xs
-
 (** Check that samplers can build and run. *)
 let () =
-  Random.init 23872;
+  Zar.self_init ();
 
+  (* Bit stream.*)
   print_endline "====== bits";
-  print_endline @@ String.trim @@ string_of_list " " string_of_bool @@
-    take_list 10 @@ Zar.bits ();
+  Zar.bits () |> Seq.take 10 |> Seq.length |> string_of_int |> print_endline;
 
   (* Coin. *)
   print_endline "\n====== coin";
-  let coin = Zar.coin 2 3 in
-  print_endline @@ string_of_bool @@ first coin;
-  print_endline @@ string_of_bool @@ first (rest coin);
-  print_endline @@ string_of_bool @@ first (rest (rest coin));
-  print_endline @@ string_of_bool @@ first (rest (rest (rest coin)));
-  print_endline @@ String.trim @@ string_of_list " " string_of_bool @@ take_list 10 coin;
+  Zar.coin_stream 2 3 |> Seq.take 10 |> Seq.length |> string_of_int |> print_endline;
 
   (* Die. *)
   print_endline "\n====== die";
-  let die = Zar.die 100 in
-  print_endline @@ string_of_int @@ first die;
-  print_endline @@ String.trim @@ string_of_list " " string_of_int @@ take_list 10 die;
+  Zar.die_stream 100 |> Seq.take 10 |> Seq.length |> string_of_int |> print_endline;
 
   (* Findist. *)
   print_endline "\n====== findist";
-  let findist = Zar.findist [1; 2; 3; 4; 5] in
-  print_endline @@ string_of_int @@ first findist;
-  print_endline @@ String.trim @@ string_of_list " " string_of_int @@ take_list 10 findist
+  Zar.findist_stream [1; 2; 3; 4; 5] |> Seq.take 10 |> Seq.length |>
+    string_of_int |> print_endline;
+
+  (* Stateful sampler. *)
+  print_endline "\n====== die sampler";
+  let sampler = Zar.die 200 in
+  List.init 10 (fun _ -> sampler#gen ()) |> List.length |>
+    string_of_int |> print_endline