+++ CHANGES.md
@@ -1,3 +1,9 @@
+## 0.9.3 (2023-04-10)
+
+* Use Seq.t instead of stream type, memoize default bit streams,
+  change init/seed functions, and include stateful 'sampler'
+  interface.
+
 ## 0.9.2 (2023-04-10)
 
 * API change: include Stream module in root Zar module.